Make the rust namespace separator work in options.
Without this flag, options are passed in a way where colon is treated as a separator by protoc. By
passing options as separate flags, this is no longer the case. The rust namespace operator is ::,
and would be understood as a separator by protoc. Using separate options flags allows passing type
attributes such as `#[derive(serde_derive::Deserialize)]`.